Refactor SPN token flow into multiple files
