+++ pkg/pop/msal.go
@@ -46,8 +46,9 @@ func AcquirePoPTokenInteractive(
 	return result.AccessToken, result.ExpiresOn.Unix(), nil
 }
 
-// acquires a PoP token using MSAL's confidential login flow. This flow does not require user interaction
-// as the credentials for the request have already been provided
+// AcquirePoPTokenConfidential acquires a PoP token using MSAL's confidential login flow.
+// This flow does not require user interaction as the credentials for the request have
+// already been provided
 func AcquirePoPTokenConfidential(
 	context context.Context,
 	popClaims map[string]string,
