+++ pkg/token/serviceprincipaltokencertificate_test.go
@@ -1,0 +1,30 @@
+package token
+
+import (
+	"testing"
+)
+
+func TestMissingCertFile(t *testing.T) {
+	p := &servicePrincipalToken{
+		clientCert: "testdata/noCertHere.pfx",
+	}
+	expectedErr := "failed to read the certificate file"
+
+	_, err := p.Token()
+	if !ErrorContains(err, expectedErr) {
+		t.Errorf("expected error %s, but got %s", expectedErr, err)
+	}
+}
+
+func TestBadCertPassword(t *testing.T) {
+	p := &servicePrincipalToken{
+		clientCert:         "testdata/testCert.pfx",
+		clientCertPassword: badSecret,
+	}
+	expectedErr := "failed to decode pkcs12 certificate while creating spt: pkcs12: decryption password incorrect"
+
+	_, err := p.Token()
+	if !ErrorContains(err, expectedErr) {
+		t.Errorf("expected error %s, but got %s", expectedErr, err)
+	}
+}
